wip backtest

+++ src/main/kotlin/org/msi/ftx1/business/Percent.kt
@@ -1,6 +1,6 @@
 package org.msi.ftx1.business
 
-class Percent(val value: Float) : Comparable<Percent> {
+class Percent(val value: Double) : Comparable<Percent> {
     infix operator fun plus(other: Percent) = value + other.value
     infix operator fun minus(other: Percent) = value - other.value
     infix operator fun div(other: Percent) = value / other.value
